Split out MFEMSubMeshBase from MFEMDomainSubMesh
