+++ framework/include/mfem/problem/MFEMProblem.h
@@ -8,8 +8,8 @@
 #include "ExternalProblem.h"
 #include "MFEMProblemData.h"
 #include "MFEMMesh.h"
-#include "MFEMSubMesh.h"
 #include "MFEMFunctorMaterial.h"
+#include "MFEMSubMeshBase.h"
 #include "MFEMVariable.h"
 #include "MFEMBoundaryCondition.h"
 #include "MFEMKernel.h"
