+++ framework/src/mfem/meshdivisions/MFEMDomainSubMesh.C
@@ -1,26 +1,22 @@
 #ifdef MFEM_ENABLED
 
-#include "MFEMSubMesh.h"
+#include "MFEMDomainSubMesh.h"
 #include "MFEMProblem.h"
 
-registerMooseObject("MooseApp", MFEMSubMesh);
+registerMooseObject("MooseApp", MFEMDomainSubMesh);
 
 InputParameters
-MFEMSubMesh::validParams()
+MFEMDomainSubMesh::validParams()
 {
-  InputParameters params = MFEMGeneralUserObject::validParams();
-  params.registerBase("MFEMSubMesh");
+  InputParameters params = MFEMSubMeshBase::validParams();
   params.addParam<std::vector<SubdomainName>>("block",
     {},
     "The list of blocks (ids) comprising this SubMesh");  
-  MooseEnum ordering("NODES VDIM", "VDIM", false);
-  params.addParam<MooseEnum>("ordering", ordering, "Ordering style to use for vector DoFs.");
-  params.addParam<int>("vdim", 1, "The number of degrees of freedom per basis function.");
   return params;
 }
 
-MFEMSubMesh::MFEMSubMesh(const InputParameters & parameters)
-  : MFEMGeneralUserObject(parameters),
+MFEMDomainSubMesh::MFEMDomainSubMesh(const InputParameters & parameters)
+  : MFEMSubMeshBase(parameters),
   _subdomain_names(getParam<std::vector<SubdomainName>>("block")),
   _subdomain_attributes(_subdomain_names.size())
 {
@@ -31,7 +27,7 @@ MFEMSubMesh::MFEMSubMesh(const InputParameters & parameters)
 }
 
 void
-MFEMSubMesh::buildSubMesh()
+MFEMDomainSubMesh::buildSubMesh()
 {
   _submesh = std::make_shared<mfem::ParSubMesh>(
   mfem::ParSubMesh::CreateFromDomain(getMFEMProblem().mesh().getMFEMParMesh(), getSubdomains()));
